Restructure PGET's code to isolate concerns
main.go -> main function and cmd setup
download/download.go -> functions for downloading data
extract/extract.go -> extract / post process functions

This includes some fixes to logic around cobra.Command and Viper
