+++ extract/extract.go
@@ -1,0 +1,63 @@
+package extract
+
+import (
+	"archive/tar"
+	"bytes"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+	"time"
+
+	"github.com/dustin/go-humanize"
+)
+
+func FromTar(buffer *bytes.Buffer, destDir string, fileSize int64) error {
+	startTime := time.Now()
+	tarReader := tar.NewReader(buffer)
+
+	for {
+		header, err := tarReader.Next()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			return err
+		}
+
+		target := filepath.Join(destDir, header.Name)
+		targetDir := filepath.Dir(target)
+		if err := os.MkdirAll(targetDir, 0755); err != nil {
+			return err
+		}
+
+		switch header.Typeflag {
+		case tar.TypeDir:
+			if err := os.MkdirAll(target, os.FileMode(header.Mode)); err != nil {
+				return err
+			}
+		case tar.TypeReg:
+			targetFile, err := os.OpenFile(target, os.O_CREATE|os.O_WRONLY, os.FileMode(header.Mode))
+			if err != nil {
+				return err
+			}
+			if _, err := io.Copy(targetFile, tarReader); err != nil {
+				targetFile.Close()
+				return err
+			}
+			targetFile.Close()
+		case tar.TypeSymlink:
+			if err := os.Symlink(header.Linkname, target); err != nil {
+				return err
+			}
+		default:
+			return fmt.Errorf("unsupported file type for %s, typeflag %s", header.Name, string(header.Typeflag))
+		}
+	}
+	elapsed := time.Since(startTime).Seconds()
+	size := humanize.Bytes(uint64(fileSize))
+	throughput := humanize.Bytes(uint64(float64(fileSize) / elapsed))
+	fmt.Printf("Extracted %s in %.3fs (%s/s)\n", size, elapsed, throughput)
+
+	return nil
+}
